create dedicated schemas folder and re-export types
