+++ src/types/expected-outcome.ts
@@ -1,145 +1,15 @@
-import { z } from 'zod';
-
-const nonEmptyString = z.string().trim().min(1);
-const optionalPositiveInt = z.number().int().positive().optional();
-const optionalString = z.string().optional();
-const optionalBoolean = z.boolean().optional();
-const selectOptionsSchema = z.array(nonEmptyString).min(1);
-
-const defaultExpectedOutcomeBaseSchema = z.object({
-  label: nonEmptyString,
-  required: optionalBoolean,
-  placeholder: optionalString,
-});
-
-const createDefaultExpectedOutcomeFieldSchemas = (
-  baseSchema: typeof defaultExpectedOutcomeBaseSchema,
-) => ({
-  text: baseSchema.extend({
-    type: z.literal('text'),
-  }),
-  textarea: baseSchema.extend({
-    type: z.literal('textarea'),
-    rows: optionalPositiveInt,
-  }),
-  chipsInput: baseSchema.extend({
-    type: z.literal('chips-input'),
-  }),
-  select: baseSchema.extend({
-    type: z.literal('select'),
-    options: selectOptionsSchema,
-  }),
-});
-
-function hasDuplicateChips(values: string[]): boolean {
-  const seen = new Set<string>();
-  for (const value of values) {
-    const normalized = value.trim().toLowerCase();
-    if (seen.has(normalized)) {
-      return true;
-    }
-    seen.add(normalized);
-  }
-  return false;
-}
-
-const defaultFieldDefinitions =
-  createDefaultExpectedOutcomeFieldSchemas(defaultExpectedOutcomeBaseSchema);
-
-export const expectedOutcomeSchemaFieldSchema = z.discriminatedUnion('type', [
-  defaultFieldDefinitions.text,
-  defaultFieldDefinitions.textarea,
-  defaultFieldDefinitions.chipsInput,
-  defaultFieldDefinitions.select,
-]);
-
-export const expectedOutcomeSchemaSchema = z
-  .array(expectedOutcomeSchemaFieldSchema)
-  .min(1);
-
-export const expectedOutcomeFieldSchema = z.discriminatedUnion('type', [
-  defaultFieldDefinitions.text.extend({
-    value: z.string(),
-  }),
-  defaultFieldDefinitions.textarea.extend({
-    value: z.string(),
-  }),
-  defaultFieldDefinitions.chipsInput.extend({
-    value: z.array(z.string()).superRefine((values, ctx) => {
-      if (hasDuplicateChips(values)) {
-        ctx.addIssue({
-          code: 'custom',
-          message:
-            'chips-input values must be unique (case-insensitive, trimmed).',
-        });
-      }
-    }),
-  }),
-  defaultFieldDefinitions.select.extend({
-    value: z.string(),
-  }),
-]);
-
-export const expectedOutcomeArraySchema = z.array(expectedOutcomeFieldSchema).min(1);
-
-export type ExpectedOutcomeSchemaField = z.infer<
-  typeof expectedOutcomeSchemaFieldSchema
->;
-export type ExpectedOutcomeSchema = z.infer<typeof expectedOutcomeSchemaSchema>;
-export type ExpectedOutcomeField = z.infer<typeof expectedOutcomeFieldSchema>;
-export type ExpectedOutcomeFieldType = ExpectedOutcomeField['type'];
-export type ExpectedOutcomeBase = z.infer<typeof defaultExpectedOutcomeBaseSchema>;
-
-export type TextExpectedOutcomeSchemaField = Extract<
+export type {
   ExpectedOutcomeSchemaField,
-  { type: 'text' }
->;
-export type TextareaExpectedOutcomeSchemaField = Extract<
-  ExpectedOutcomeSchemaField,
-  { type: 'textarea' }
->;
-export type ChipsExpectedOutcomeSchemaField = Extract<
-  ExpectedOutcomeSchemaField,
-  { type: 'chips-input' }
->;
-export type SelectExpectedOutcomeSchemaField = Extract<
-  ExpectedOutcomeSchemaField,
-  { type: 'select' }
->;
-
-export type TextExpectedOutcomeField = Extract<
+  ExpectedOutcomeSchema,
   ExpectedOutcomeField,
-  { type: 'text' }
->;
-export type TextareaExpectedOutcomeField = Extract<
-  ExpectedOutcomeField,
-  { type: 'textarea' }
->;
-export type ChipsExpectedOutcomeField = Extract<
-  ExpectedOutcomeField,
-  { type: 'chips-input' }
->;
-export type SelectExpectedOutcomeField = Extract<
-  ExpectedOutcomeField,
-  { type: 'select' }
->;
-
-export function validateExpectedOutcomeSchema(
-  schema: unknown,
-): asserts schema is ExpectedOutcomeSchema {
-  const parsed = expectedOutcomeSchemaSchema.safeParse(schema);
-  if (!parsed.success) {
-    throw new Error(
-      `Invalid expectedOutcomeSchema: ${parsed.error.issues[0].message}`,
-    );
-  }
-}
-
-export function validateExpectedOutcomeArray(
-  expectedOutcome: unknown,
-): asserts expectedOutcome is ExpectedOutcomeField[] {
-  const parsed = expectedOutcomeArraySchema.safeParse(expectedOutcome);
-  if (!parsed.success) {
-    throw new Error(`Invalid expectedOutcome: ${parsed.error.issues[0].message}`);
-  }
-}
+  ExpectedOutcomeFieldType,
+  ExpectedOutcomeBase,
+  TextExpectedOutcomeSchemaField,
+  TextareaExpectedOutcomeSchemaField,
+  ChipsExpectedOutcomeSchemaField,
+  SelectExpectedOutcomeSchemaField,
+  TextExpectedOutcomeField,
+  TextareaExpectedOutcomeField,
+  ChipsExpectedOutcomeField,
+  SelectExpectedOutcomeField,
+} from '../schemas/expected-outcome';
